Added code to get product data in ThingListFragment.
Added network code to get product info in ProductFetcher.

Need to adjust it to JSON format in Outpan and use barcode and name intially to set fields when scanning a new item.

Finally need to make the network robust by checking for connection and make a preference activity page to control network usage.