Added code to get product data in ThingListFragment.
Added network code to get product info in ProductFetcher.

Need to adjust it to JSON format in Outpan and use barcode and name intially to set fields when scanning a new item.

Finally need to make the network robust by checking for connection and make a preference activity page to control network usage.

////

Added class to detect a connection to internet.
Need to refactor robust networking code to ThingFragment from ThingListFragment

Added utility classes to monitor and control network usage